adds Charger Card component to SearchLocation page
