+++ src/context/theme.js
@@ -12,6 +12,7 @@ export const theme = createTheme({
     Card: {
       maxWidth: 350,
       margin: 'auto',
+      background: '#e0e0a3',
     },
     background: {
       card: '#e0e0e080',
